requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
matplotlib>=3.7
lightgbm>=4.0
torch>=2.1
tqdm>=4.65
tabulate>=0.9
